Få periode i egen komponent
